Attackspeed lehet double is #4

+++ Main.java
@@ -5,12 +5,12 @@ public class Main {
         // Paraméterek bekérése Scanner-el
         Scanner sc = new Scanner(System.in);
 
-        int unit1Hp = readIn(sc, "Első egység HP-ja: ");
-        int unit1Dmg = readIn(sc, "Első egység DMG-je: ");
-        int unit1As = readIn(sc, "Első egység AS-je: ");
-        int unit2Hp = readIn(sc, "Második egység HP-ja: ");
-        int unit2Dmg = readIn(sc, "Második egység DMG-je: ");
-        int unit2As = readIn(sc, "Második egység AS-je: ");
+        double unit1Hp = readIn(sc, "Első egység HP-ja: ");
+        double unit1Dmg = readIn(sc, "Első egység DMG-je: ");
+        double unit1As = readIn(sc, "Első egység AS-je: ");
+        double unit2Hp = readIn(sc, "Második egység HP-ja: ");
+        double unit2Dmg = readIn(sc, "Második egység DMG-je: ");
+        double unit2As = readIn(sc, "Második egység AS-je: ");
 
         // Két Unit létrehozása
         Unit unit1 = new Unit("Harcos", unit1Dmg, unit1Hp, unit1As);
@@ -22,13 +22,13 @@ public class Main {
         battle(unit1, unit2);
     }
 
-    private static int readIn(Scanner sc, String msg) {
+    private static double readIn(Scanner sc, String msg) {
         boolean acceptable = true;
-        int answer = 0;
+        double answer = 0;
         do {
             System.out.print(msg);
-            if (sc.hasNextInt()) {
-                answer = sc.nextInt();
+            if (sc.hasNextFloat()) {
+                answer = sc.nextFloat();
                 acceptable = false;
             } else {
                 System.out.println(sc.next() + " nem érvényes értéket adott meg, adjon meg egy számot! ");
@@ -38,8 +38,8 @@ public class Main {
     }
 
     private static void battle(Unit unit1, Unit unit2) {
-        int as1 = unit1.getAs();
-        int as2 = unit2.getAs();
+        double as1 = unit1.getAs();
+        double as2 = unit2.getAs();
         String name1 = unit1.getName();
         String name2 = unit2.getName();
         boolean bothAlive = true;
